requirements.txt
anthropic>=0.18.0
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
resend>=0.8.0
markdown>=3.5.0
feedparser>=6.0.0
